fix #38
Reorganization of builtin folder.
Juste add a check for unset, to print a message if the are specials
character. Need to create a special error management for export, i cn't
use this."
