--- srcs/execution/builtins/echo_utils.c
@@ -1,18 +1,0 @@
-#include "execution.h"
-
-int	check_n(char *arg)
-{
-	int	i;
-
-	i = 1;
-	if (arg[0] != '-')
-		return (-1);
-	while (arg[i])
-	{
-		if (arg[i] == 'n')
-			i++;
-		else
-			return (-1);
-	}
-	return (0);
-}
